BallShoot Game
The ball shoot game is a game in java in which one has to click on the
balls on the screen to score.
There are 3 buttons : Start, Stop and Pause.
The start button is used to start the game or to add more balls, stop
button quits the current game and displays the score and pause button
is used to pause/restart the game.